refactor(popups): promote shared <TokenGrid> from ClientPopup
ClientPopup already had a local TokenGrid (paged tile grid + arrows);
Contact/ProfileSet inlined the byte-identical block and TokenPopup's
SuperToken branch a `Pagination half` variant.  Hoist it into
perpShared and reuse in all four — the paged-grid layout (the
mobile-port surface) now lives once.  Open-state/subpop overlay stay
in the callers (the share-markup / keep-behaviour split from
<PopupHeader>/<PopupMenu>).  Callers drop their now-encapsulated
page/pageCount/pageTokens state.
