+++ scripts/components/popups/ProfileSetPopup.tsx
@@ -9,7 +9,7 @@ import { useState } from 'preact/hooks';
 import type { ProfileSetPopupVM } from '../../game/profilesetView.js';
 import { PopupHeader } from './PopupHeader.js';
 import type { PreactDialogHandle } from './dialogManager.js';
-import { PageArrows, TokenSubpop, TokenTile, fireAction } from './perpShared.js';
+import { TokenGrid, TokenSubpop, fireAction } from './perpShared.js';
 
 export interface ProfileSetPopupProps {
   vm: ProfileSetPopupVM;
@@ -20,11 +20,7 @@ export interface ProfileSetPopupProps {
 }
 
 export function ProfileSetPopup({ vm, onClose, popup }: ProfileSetPopupProps) {
-  const [page, setPage] = useState(0);
   const [openToken, setOpenToken] = useState<string | null>(null);
-
-  const pageCount = Math.max(1, Math.ceil(vm.tokens.length / vm.pageSize));
-  const pageTokens = vm.tokens.slice(page * vm.pageSize, page * vm.pageSize + vm.pageSize);
 
   return (
     <div class="PopupBody">
@@ -47,18 +43,13 @@ export function ProfileSetPopup({ vm, onClose, popup }: ProfileSetPopupProps) {
               />
             ))}
           </div>
-          <div class="Pagination">
-            <div class="PopupTokens">
-              <div class="PopupPageWrap">
-                <div class="PopupPage" data-page-id={page}>
-                  {pageTokens.map((t) => (
-                    <TokenTile key={t.gestalt} token={t} onOpen={setOpenToken} />
-                  ))}
-                </div>
-              </div>
-            </div>
-            <PageArrows page={page} pageCount={pageCount} setPage={setPage} />
-          </div>
+          <TokenGrid
+            tokens={vm.tokens}
+            pageSize={vm.pageSize}
+            paginationClass="Pagination"
+            tokensClass="PopupTokens"
+            onOpen={setOpenToken}
+          />
           <div class="PopupSummary">
             <div class="PopupSummaryItem Profiles">
               <div class="RenderSprite Tobi" />
