+++ scripts/components/popups/perpShared.tsx
@@ -6,6 +6,7 @@
 // once here instead of being duplicated per perp component.
 
 import type { JSX } from 'preact';
+import { useState } from 'preact/hooks';
 import type { ProvidedSubpopVM, ProvidedTileVM } from '../../game/providedView.js';
 import type { TokenUpgradeSubpopVM } from '../../game/tokenPopupView.js';
 import type { TokenVM } from '../../game/tokenView.js';
@@ -129,6 +130,43 @@ export function PageArrows({
         onClick={() => setPage((p) => Math.max(p - 1, 0))}
       />
     </>
+  );
+}
+
+/** One paged token grid — `token.html` inside `profileset*.html`.
+ *  Owns its own page state; the wrapper classes + page size vary per
+ *  caller (Contact/ProfileSet `Pagination`/`PopupTokens`, Client's
+ *  provided/consumed halves, TokenPopup's SuperToken `Pagination
+ *  half`).  The subpop overlay + open-state stay with the caller. */
+export function TokenGrid({
+  tokens,
+  pageSize,
+  paginationClass,
+  tokensClass,
+  onOpen,
+}: {
+  tokens: TokenVM[];
+  pageSize: number;
+  paginationClass: string;
+  tokensClass: string;
+  onOpen: (gestalt: string) => void;
+}) {
+  const [page, setPage] = useState(0);
+  const pageCount = Math.max(1, Math.ceil(tokens.length / pageSize));
+  const pageTokens = tokens.slice(page * pageSize, page * pageSize + pageSize);
+  return (
+    <div class={paginationClass}>
+      <div class={tokensClass}>
+        <div class="PopupPageWrap">
+          <div class="PopupPage" data-page-id={page}>
+            {pageTokens.map((t) => (
+              <TokenTile key={t.gestalt} token={t} onOpen={onOpen} />
+            ))}
+          </div>
+        </div>
+      </div>
+      <PageArrows page={page} pageCount={pageCount} setPage={setPage} />
+    </div>
   );
 }
